Pull php code out Appt_Enc reports to ApptEncouterController.
