--- interface/reports/amc_tracking.php
@@ -1,403 +1,0 @@
-<?php
-/*
- *
- * Copyright (C) 2016-2017 Terry Hill <teryhill@librehealth.io> 
- * Copyright (C) 2011 Brady Miller <brady@sparmy.com>
- *
- * LICENSE: This program is free software; you can redistribute it and/or 
- * modify it under the terms of the GNU General Public License 
- * as published by the Free Software Foundation; either version 3 
- * of the License, or (at your option) any later version. 
- * This program is distributed in the hope that it will be useful, 
- * but WITHOUT ANY WARRANTY; without even the implied warranty of 
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the 
- * GNU General Public License for more details. 
- * You should have received a copy of the GNU General Public License 
- * along with this program. If not, see <http://opensource.org/licenses/gpl-license.php>;. 
- *
- * LICENSE: This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
- * See the Mozilla Public License for more details.
- * If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
- * 
- * @package LibreHealth EHR 
- * @author Brady Miller <brady@sparmy.com> 
- * @link http://librehealth.io 
- */
-
-//SANITIZE ALL ESCAPES
-$sanitize_all_escapes=true;
-//
-
-//STOP FAKE REGISTER GLOBALS
-$fake_register_globals=false;
-//
-
-require_once("../globals.php");
-require_once("../../library/patient.inc");
-require_once("../../library/report_functions.php");
-require_once("$srcdir/formatting.inc.php");
-require_once "$srcdir/options.inc.php";
-require_once "$srcdir/formdata.inc.php";
-require_once "$srcdir/amc.php";
-$DateFormat = DateFormatRead(true);
-$DateLocale = getLocaleCodeForDisplayLanguage($GLOBALS['language_default']);
-
-// Collect form parameters (set defaults if empty)
-$form_from_date = (isset($_POST['form_from_date'])) ? trim($_POST['form_from_date']) : "";
-$form_to_date = (isset($_POST['form_to_date'])) ? trim($_POST['form_to_date']) : "";
-$rule = (isset($_POST['form_rule'])) ? trim($_POST['form_rule']) : "";
-$provider  = trim($_POST['form_provider']);
-
-?>
-
-<html>
-
-<head>
-<?php html_header_show();?>
-
-<link rel="stylesheet" href="<?php echo $css_header;?>" type="text/css">
-
-<title><?php echo htmlspecialchars( xl('Automated Measure Calculations (AMC) Tracking'), ENT_NOQUOTES); ?></title>
-<link rel="stylesheet" href="../../library/css/jquery.datetimepicker.css">
-
-<script type="text/javascript" src="../../library/overlib_mini.js"></script>
-<script type="text/javascript" src="../../library/textformat.js"></script>
-<script type="text/javascript" src="../../library/dialog.js"></script>
-<script type="text/javascript" src="../../library/js/jquery-1.9.1.min.js"></script>
-
-<script LANGUAGE="JavaScript">
-
- var mypcc = '<?php echo $GLOBALS['phone_country_code'] ?>';
-
- $(document).ready(function() {
-  var win = top.printLogSetup ? top : opener.top;
-  win.printLogSetup(document.getElementById('printbutton'));
- });
-
- function send_sum(patient_id,transaction_id) {
-   if ( $('#send_sum_flag_' + patient_id + '_' + transaction_id).attr('checked') ) {
-     var mode = "add";
-   }
-   else {
-     var mode = "remove";
-   }
-   top.restoreSession();
-   $.post( "../../library/ajax/amc_misc_data.php",
-     { amc_id: "send_sum_amc",
-       complete: true,
-       mode: mode,
-       patient_id: patient_id,
-       object_category: "transactions",
-       object_id: transaction_id
-     }
-   );
- }
-
- function send_sum_elec(patient_id,transaction_id) {
-   if ( $('#send_sum_elec_flag_' + patient_id + '_' + transaction_id).attr('checked') ) {
-     if ( !$('#send_sum_flag_' + patient_id + '_' + transaction_id).attr('checked') ) {
-       $('#send_sum_elec_flag_' + patient_id + '_' + transaction_id).removeAttr("checked");  
-       alert("<?php echo xls('Can not set this unless the Summary of Care Sent toggle is set.'); ?>");
-       return false;
-     }
-     var mode = "add";
-   }
-   else {
-     var mode = "remove";
-   }
-   top.restoreSession();
-   $.post( "../../library/ajax/amc_misc_data.php",
-     { amc_id: "send_sum_elec_amc",
-       complete: true,
-       mode: mode,
-       patient_id: patient_id,
-       object_category: "transactions",
-       object_id: transaction_id
-     }
-   );
- }
-
- function provide_rec_pat(patient_id,date_created) {
-   if ( $('#provide_rec_pat_flag_' + patient_id ).attr('checked') ) {
-     var mode = "complete_safe";
-   }
-   else {
-     var mode = "uncomplete_safe";
-   }
-   top.restoreSession();
-   $.post( "../../library/ajax/amc_misc_data.php",
-     { amc_id: "provide_rec_pat_amc",
-       complete: true,
-       mode: mode,
-       date_created: date_created,
-       patient_id: patient_id
-     }
-   );
- }
-
- function provide_sum_pat(patient_id,encounter_id) {
-   if ( $('#provide_sum_pat_flag_' + patient_id + '_' + encounter_id).attr('checked') ) {
-     var mode = "add";
-   }
-   else {
-     var mode = "remove";
-   }
-   top.restoreSession();
-   $.post( "../../library/ajax/amc_misc_data.php",
-     { amc_id: "provide_sum_pat_amc",
-       complete: true,
-       mode: mode,
-       patient_id: patient_id,
-       object_category: "form_encounter",
-       object_id: encounter_id
-     }
-   );
- }
-
-</script>
-
-<style type="text/css">
-
-/* specifically include & exclude from printing */
-@media print {
-    #report_parameters {
-        visibility: hidden;
-        display: none;
-    }
-    #report_parameters_daterange {
-        visibility: visible;
-        display: inline;
-    }
-    #report_results table {
-       margin-top: 0px;
-    }
-}
-
-/* specifically exclude some from the screen */
-@media screen {
-    #report_parameters_daterange {
-        visibility: hidden;
-        display: none;
-    }
-}
-
-</style>
-</head>
-
-<body class="body_top">
-
-<!-- Required for the popup date selectors -->
-<div id="overDiv" style="position:absolute; visibility:hidden; z-index:1000;"></div>
-
-<span class='title'><?php echo htmlspecialchars( xl('Report'), ENT_NOQUOTES); ?> - 
-
-<?php echo htmlspecialchars( xl('Automated Measure Calculations (AMC) Tracking'), ENT_NOQUOTES); ?></span>
-
-<form method='post' name='theform' id='theform' action='amc_tracking.php' onsubmit='return top.restoreSession()'>
-
-<div id="report_parameters">
-
-<table>
- <tr>
-  <td width='470px'>
-    <div style='float:left'>
-
-    <table class='text'>
-
-                 <?php showFromAndToDates(); ?>
-
-                <tr>
-                        <td class='label'>
-                            <?php echo htmlspecialchars( xl('Rule'), ENT_NOQUOTES); ?>:
-                        </td>
-                        <td>
-                            <select name='form_rule'>
-                            <option value='send_sum_amc' <?php if ($rule == "send_sum_amc") echo "selected"; ?>>
-                            <?php echo htmlspecialchars( xl('Send Summaries with Referrals'), ENT_NOQUOTES); ?></option>
-                            <option value='provide_rec_pat_amc' <?php if ($rule == "provide_rec_pat_amc") echo "selected"; ?>>
-                            <?php echo htmlspecialchars( xl('Patient Requested Medical Records'), ENT_NOQUOTES); ?></option>
-                            <option value='provide_sum_pat_amc' <?php if ($rule == "provide_sum_pat_amc") echo "selected"; ?>>
-                            <?php echo htmlspecialchars( xl('Provide Records to Patient for Visit'), ENT_NOQUOTES); ?></option>
-                            </select>
-                        </td>
-                </tr>
-
-                <tr>      
-            <td class='label'>
-               <?php echo htmlspecialchars( xl('Provider'), ENT_NOQUOTES); ?>:
-            </td>
-            <td>
-              <?php // Build a drop-down list of providers.
-                dropDownProviders();
-              ?>
-            </td>
-        </tr>
-    </table>
-
-    </div>
-
-  </td>
-  <td align='left' valign='middle' height="100%">
-    <table style='border-left:1px solid; width:100%; height:100%' >
-        <tr>
-            <td>
-                <div style='margin-left:15px'>
-                    <a href='#' class='css_button' onclick='$("#form_refresh").attr("value","true"); top.restoreSession(); $("#theform").submit();'>
-                    <span>
-                        <?php echo htmlspecialchars( xl('Submit'), ENT_NOQUOTES); ?>
-                    </span>
-                    </a>
-                                        <?php if ($_POST['form_refresh']) { ?>
-                    <a href='#' class='css_button' id='printbutton'>
-                        <span>
-                            <?php echo htmlspecialchars( xl('Print'), ENT_NOQUOTES); ?>
-                        </span>
-                    </a>
-                    <?php } ?>
-                </div>
-            </td>
-        </tr>
-    </table>
-  </td>
- </tr>
-</table>
-
-</div>  <!-- end of search parameters -->
-
-<br>
-
-<?php
- if ($_POST['form_refresh']) {
-?>
-
-
-<div id="report_results">
-<table>
-
- <thead>
-  <th>
-   <?php echo htmlspecialchars( xl('Patient Name'), ENT_NOQUOTES); ?>
-  </th>
-
-  <th>
-   <?php echo htmlspecialchars( xl('Patient ID'), ENT_NOQUOTES); ?>
-  </th>
-
-  <th>
-    <?php
-      if ($rule == "send_sum_amc") {
-        echo htmlspecialchars( xl('Referral Date'), ENT_NOQUOTES);
-      }
-      else if ($rule == "provide_rec_pat_amc") {
-        echo htmlspecialchars( xl('Record Request Date'), ENT_NOQUOTES);
-      }
-      else { // $rule == "provide_sum_pat_amc"
-        echo htmlspecialchars( xl('Encounter Date'), ENT_NOQUOTES);
-      }
-    ?>
-  </th>
-
-  <th>
-   <?php
-     if ($rule == "send_sum_amc") {
-       echo htmlspecialchars( xl('Referral ID'), ENT_NOQUOTES);
-     }
-      else if ($rule == "provide_rec_pat_amc") {
-        echo "&nbsp";
-      }
-     else { // $rule == "provide_sum_pat_amc"
-       echo htmlspecialchars( xl('Encounter ID'), ENT_NOQUOTES);
-     }
-   ?>
-  </th>
-
-  <th>
-   <?php 
-     if ($rule == "provide_rec_pat_amc") {
-       echo htmlspecialchars( xl('Medical Records Sent'), ENT_NOQUOTES);
-     }
-     else if ($rule == "send_sum_amc") {
-       echo htmlspecialchars( xl('Summary of Care Sent'), ENT_NOQUOTES);
-     }
-     else { // $rule == "provide_sum_pat_amc"
-       echo htmlspecialchars( xl('Medical Summary Given'), ENT_NOQUOTES);
-     }
-   ?>
-  </th>
-  <?php
-    if ($rule == "send_sum_amc") {
-      echo "<th>";
-      echo htmlspecialchars( xl('Summary of Care Sent Electronically'), ENT_NOQUOTES);
-      echo "<th>";
-    }
-  ?>
-
- </thead>
- <tbody>  <!-- added for better print-ability -->
-<?php
-
-  // Send the request for information
-  $resultsArray = amcTrackingRequest($rule,$form_from_date,$form_to_date,$provider);
-
-?>
-
-  <?php
-   foreach ($resultsArray as $result) {
-     echo "<tr bgcolor='" . $bgcolor ."'>";
-     echo "<td>" . htmlspecialchars($result['lname'].",".$result['fname'], ENT_NOQUOTES) . "</td>";
-     echo "<td>" . htmlspecialchars($result['pid'],ENT_NOQUOTES) . "</td>";
-     echo "<td>" . date($DateFormat, strtotime(attr($result['date']))) . "</td>";
-     if ($rule == "send_sum_amc" || $rule == "provide_sum_pat_amc") {
-       echo "<td>" . htmlspecialchars($result['id'],ENT_NOQUOTES) . "</td>";
-     }
-     else { //$rule == "provide_rec_pat_amc"
-       echo "&nbsp";
-     }
-
-     if ($rule == "send_sum_amc") {
-       echo "<td><input type='checkbox' id='send_sum_flag_".attr($result['pid'])."_".attr($result['id'])."' onclick='send_sum(\"".htmlspecialchars($result['pid'],ENT_QUOTES)."\",\"".htmlspecialchars($result['id'],ENT_QUOTES)."\")'>" . htmlspecialchars( xl('Yes'), ENT_NOQUOTES) . "</td>";
-       echo "<td><input type='checkbox' id='send_sum_elec_flag_".attr($result['pid'])."_".attr($result['id'])."' onclick='send_sum_elec(\"".htmlspecialchars($result['pid'],ENT_QUOTES)."\",\"".htmlspecialchars($result['id'],ENT_QUOTES)."\")'>" . htmlspecialchars( xl('Yes'), ENT_NOQUOTES) . "</td>";
-     }
-     else if ($rule == "provide_rec_pat_amc") {
-       echo "<td><input type='checkbox' id='provide_rec_pat_flag_".attr($result['pid'])."' onclick='provide_rec_pat(\"".htmlspecialchars($result['pid'],ENT_QUOTES)."\",\"".htmlspecialchars($result['date'],ENT_QUOTES)."\")'>" . htmlspecialchars( xl('Yes'), ENT_NOQUOTES) . "</td>";
-     }
-     else { //$rule == "provide_sum_pat_amc"
-       echo "<td><input type='checkbox' id='provide_sum_pat_flag_".attr($result['pid'])."_".attr($result['id'])."' onclick='provide_sum_pat(\"".htmlspecialchars($result['pid'],ENT_QUOTES)."\",\"".htmlspecialchars($result['id'],ENT_QUOTES)."\")'>" . htmlspecialchars( xl('Yes'), ENT_NOQUOTES) . "</td>";
-     }
-     echo "</tr>";
-   }
-  ?>
-
-</tbody>
-</table>
-</div>  <!-- end of search results -->
-<?php } else { ?>
-<div class='text'>
-    <?php echo htmlspecialchars( xl('Please input search criteria above, and click Submit to view results.'), ENT_NOQUOTES); ?>
-</div>
-<?php } ?>
-
-<input type='hidden' name='form_refresh' id='form_refresh' value=''/>
-
-</form>
-
-</body>
-
-<script type="text/javascript" src="../../library/js/jquery.datetimepicker.full.min.js"></script>
-<script>
-    $(function() {
-        $("#form_from_date").datetimepicker({
-            timepicker: true,
-            format: "<?= $DateFormat; ?>"
-        });
-        $("#form_to_date").datetimepicker({
-            timepicker: true,
-            format: "<?= $DateFormat; ?>"
-        });
-        $.datetimepicker.setLocale('<?= $DateLocale;?>');
-    });
-
-</script>
-
-</html>
-
